Consolidate transport resource registration logic.
