+++ clustering/infinispan/extension/src/main/java/org/jboss/as/clustering/infinispan/subsystem/NoTransportResourceDefinition.java
@@ -22,10 +22,7 @@
 
 package org.jboss.as.clustering.infinispan.subsystem;
 
-import org.jboss.as.clustering.controller.ManagementResourceRegistration;
-import org.jboss.as.clustering.controller.ResourceDescriptor;
-import org.jboss.as.clustering.controller.SimpleResourceRegistration;
-import org.jboss.as.clustering.controller.ResourceServiceHandler;
+import org.jboss.as.clustering.function.Consumers;
 import org.jboss.as.controller.ModelVersion;
 import org.jboss.as.controller.PathElement;
 import org.jboss.as.controller.transform.description.ResourceTransformationDescriptionBuilder;
@@ -41,17 +38,6 @@ public class NoTransportResourceDefinition extends TransportResourceDefinition {
     }
 
     NoTransportResourceDefinition() {
-        super(PATH);
-    }
-
-    @Override
-    public void register(ManagementResourceRegistration parentRegistration) {
-        ManagementResourceRegistration registration = parentRegistration.registerSubModel(this);
-
-        ResourceDescriptor descriptor = new ResourceDescriptor(this.getResourceDescriptionResolver())
-                .addCapabilities(CLUSTERING_CAPABILITIES.values())
-                ;
-        ResourceServiceHandler handler = new NoTransportServiceHandler();
-        new SimpleResourceRegistration(descriptor, handler).register(registration);
+        super(PATH, Consumers.empty(), new NoTransportServiceHandler());
     }
 }
